pyproject.toml
[project]
name = "spm2ytm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.1",
    "google-api-python-client>=2.187.0",
    "google-auth>=2.43.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.2",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "spotipy>=2.25.1",
    "yt-dlp>=2025.11.12",
    "ytmusicapi>=1.11.1",
]
